added ServerEventProvider
